\name{pwrEWAS.data}
\alias{pwrEWAS.data}
\title{pwrEWAS.data: Reference data accompanying pwrEWAS}
\description{
This package provides reference data required for the pwrEWAS package.
Additional information about the origin of each reference 
data set can be found in the vignette: vignette("pwrEWAS.data")
}
\keyword{DNAm}
\keyword{microarray}
\keyword{tissue}
