\name{pwrEWAS.data}
\alias{pwrEWAS.data}
\title{pwrEWAS_data: Reference data accompanying pwrEWAS}
\description{
This package provides reference data required for pwrEWAS.
Additional information about the origin of each reference 
data set can be found in the vignette.
}
\keyword{DNAm}
\keyword{microarray}
\keyword{tissue}
